+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changelog
 - Clear fixture finalizers when removing cached results from failed fixtures
   to fix compatibility with pytest >= 9, which asserts that ``_finalizers`` is
   empty before executing a fixture.
+  Fixes `#323 <https://github.com/pytest-dev/pytest-rerunfailures/issues/323>`_.
 
 
 16.1 (2025-10-10)